Finished FXAA tutorial

+++ Assets/Custom RP/ShaderLibrary/Lighting.hlsl
@@ -38,7 +38,7 @@ float3 GetLighting(Surface surfaceWS, BRDF brdf, GI gi)
 #if defined(_LIGHTS_PER_OBJECT)
 	for(int j = 0; j < min(unity_LightData.y, 8); j++)
 	{
-		int lightIndex = unity_LightIndices[j / 4][j % 4];
+		int lightIndex = unity_LightIndices[(uint)j / 4][(uint)j % 4];
 		Light light = GetOtherLight(lightIndex, surfaceWS, shadowData);
 		if (RenderingLayersOverlap(surfaceWS, light))
 		{
